% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braidIndex.R
\name{braidIndex}
\alias{braidIndex}
\title{Calculate morphological indexes multiple rivers or reaches}
\usage{
braidIndex(Infos, rel_path = "")
}
\arguments{
\item{Infos}{a table containing river-related details.}
}
\value{
a table of morphological indexes
}
\description{
Calculate morphological indexes multiple rivers or reaches
}
