% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braidIndex_one.R
\name{braidIndex_one}
\alias{braidIndex_one}
\title{Calculate morphological indexes for one river or reach}
\usage{
braidIndex_one(rivdata, area, points_space)
}
\arguments{
\item{rivdata}{river data}

\item{area}{total area}

\item{points_space}{space between points}
}
\value{
a 1-row table of morphological indexes
}
\description{
Calculate morphological indexes for one river or reach
}
