% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braidIndex_plot.R
\name{braidIndex_plot}
\alias{braidIndex_plot}
\title{Plot index (either W`*` or BRI`*`)}
\usage{
braidIndex_plot(
  tib_morpho_index,
  index = "W",
  position,
  color,
  threshold = 0.004
)
}
\arguments{
\item{tib_morpho_index}{the table with morphological indexes values, resulting from a call to braidIndex().}

\item{position}{the variable according to which estimates should be placed on graphic.}

\item{color}{the variable according to which estimates should be colored.}

\item{threshold}{the threshold for a section to be considered as braiding (defaults to 0.004).}
}
\value{
plot of index
}
\description{
Plot index (either W`*` or BRI`*`)
}
