% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braidIndex_read.R
\name{braidIndex_read}
\alias{braidIndex_read}
\title{Read data to}
\usage{
braidIndex_read(File)
}
\arguments{
\item{File}{must have ";" as column separators and extension ".txt". Must have columns "ID","RASTERVALU","Z","TYPO_VEGE","NAME".}
}
\value{

}
\description{
Read data to
}
\examples{
braidIndex_read("data-raw/Drac_Chabottes_2018.txt")
}
